% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSKAT.R
\name{aSKAT}
\alias{aSKAT}
\title{Small-sample adjusted SKAT for a univariate binary/Gaussian outcome ('a' stands for 'adjusted')}
\usage{
aSKAT(
  formula.H0,
  data = NULL,
  K,
  type = c("binary", "continuous"),
  acc = 1e-05,
  lim = 10000,
  tol = 1e-10
)
}
\arguments{
\item{formula.H0}{a two-sided linear formula object under the null, indicating the variables to adjust.}

\item{data}{a data frame (required) containing the variables named in formula.}

\item{K}{the kernel matrix, which quantifies the similarities between samples.}

\item{type}{the outcome type: binary or continous.}

\item{acc}{error bound for \code{davies} method.}

\item{lim}{maximum number of integration terms for \code{davies} method.}

\item{tol}{the eigenvalue cutoff, below which is considered to be 0. This is used to reduce the computation burden.}
}
\value{
\describe{A list containing
  \item{p.value}{ association p-value}
  \item{Q.adj}{ adjusted score statistic}
}
}
\description{
Compute the adjusted score statistic and p-value for a univariate bianry/Gaussian outcome.
}
\examples{
set.seed(123)
Y <- rnorm(100)
Z <- matrix(rnorm(200), 100, 2)
G <- matrix(rbinom(1000, 1, 0.25), 100, 10)
K <- G \%*\% t(G)
data <- data.frame(Y = Y, Z = Z)
aSKAT(Y ~ Z, data = data, K = K, type = 'continuous')
aSKAT(Y ~ 1, K = K, type = 'continuous')  # No covariate
Y <- rbinom(100, 1, 0.5)
data <- data.frame(Y = Y, Z = Z)
aSKAT(Y ~ Z, data = data, K = K, type = 'binary')
aSKAT(Y ~ 1, data = data, K = K, type = 'binary')    # No covariate
}
\references{
Chen et al. (2016) Small Sample Kernel Association Tests for Human Genetic and Microbiome Association Studies.
Genet Epidemiol. 40(1):5-19.
}
\author{
Jun Chen
}
\keyword{SKAT}
