% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSKAT.R
\name{cSKAT}
\alias{cSKAT}
\title{Small-sample adjusted SKAT for a  correlated (continuous) outcome ('c' stands for 'correlated').}
\usage{
cSKAT(
  formula.H0,
  data = NULL,
  K,
  relmat = NULL,
  acc = 1e-05,
  lim = 10000,
  tol = 1e-10,
  ...
)
}
\arguments{
\item{formula.H0}{a two-sided linear formula object describing both the fixed-effects and random-effects part of
the model under the null, use the same syntax as the \code{lmer} in \code{lme4} package.}

\item{data}{a data frame (required) containing the variables named in formula.}

\item{K}{the kernel matrix, which quantifies the similarities between samples.}

\item{relmat}{an optional correlation structure, see \code{lme4qtl} package.}

\item{acc}{error bound for \code{davies} method.}

\item{lim}{maximum number of integration terms for \code{davies} method.}

\item{tol}{the eigenvalue cutoff, below which it is considered to be 0. This is used to reduce the computation burden.}

\item{...}{arguments passing to mixed model \code{glmer} and \code{relmatGlmer}.}
}
\value{
A list containing
\describe{ 
  \item{p.value}{ association p-value}
  \item{Q.adj}{ adjusted score statistic}
}
}
\description{
Compute the adjusted score statistic and p-value for a correlated outcome
}
\examples{
set.seed(123)
Y <- rnorm(100)
Z <- matrix(rnorm(200), 100, 2)
ID <- gl(20, 5)
G <- matrix(rbinom(1000, 1, 0.05), 100, 10)
K <- G \%*\% t(G)
data <- data.frame(Y = Y, Z = Z, ID = ID)
cSKAT(Y ~ Z + (1 | ID), data = data, K = K)
cSKAT(Y ~ (1 | ID), data = data, K = K)  # No covariate
}
\references{
Zhan X, et al. (2018) A small-sample kernel association test for correlated data with application to microbiome association studies.  
Genet Epidemiol.;42(8):772-782
}
\author{
Jun Chen
}
\keyword{SKAT}
