% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSKAT.R
\name{mSKAT}
\alias{mSKAT}
\title{Small-sample adjusted SKAT for a multivariate (continuous) outcome ('m' stands for 'multivariate') with de-correlation. 
De-correlation is effective when the number of outcome variables is small compared to the sample size.}
\usage{
mSKAT(
  formula.H0 = NULL,
  data = NULL,
  Y,
  K,
  acc = 1e-05,
  lim = 10000,
  tol = 1e-10
)
}
\arguments{
\item{formula.H0}{a one-sided linear formula object under the null, indicating the variables to adjust.}

\item{data}{a data frame containing the variables named in formula. It is required if the formula is not null or the intercept model (~ 1).}

\item{Y}{a matrix for the multivariate outcome, row - samples, col - phenotypes.}

\item{K}{the kernel matrix, which quantifies the similarities between samples.}

\item{acc}{error bound for \code{davies} method.}

\item{lim}{maximum number of integration terms for \code{davies} method.}

\item{tol}{the eigenvalue cutoff, below which is considered to be 0. This is used to reduce the computation burden.}
}
\value{
A list containing
\describe{
  \item{p.value}{ association p-value}
  \item{Q.adj}{ adjusted score statistic}
}
}
\description{
Compute the adjusted score statistic and p-value for a multivariate continuous outcome with de-correlation.
}
\examples{
set.seed(123)
Y <- matrix(rnorm(400), 100, 4)
Z <- matrix(rnorm(200), 100, 2)
G <- matrix(rbinom(1000, 1, 0.25), 100, 10)
K <- G \%*\% t(G)
data <- data.frame(Z = Z)
mSKAT(~ Z, data = data, Y = Y, K = K) 
mSKAT(Y = Y, K = K)    # No covariate
mSKAT(~ 1, Y = Y, K = K)    # No covariate
}
\references{
Zhan X, et al. (2017)  A small-sample multivariate kernel machine test for microbiome association studies.
Genet Epidemiol. 41(3):210-220.
}
\author{
Jun Chen
}
\keyword{SKAT}
