% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSKAT.R
\name{VSAT}
\alias{VSAT}
\title{Small-sample adjusted SKAT for a generalized mixed effects model (GLMM). Since it encompasses a wide range of designs,
we name it more generally as Variant-Set Association Test (VSAT).}
\usage{
VSAT(
  formula.H0,
  data = NULL,
  K,
  relmat = NULL,
  family,
  acc = 1e-05,
  lim = 10000,
  tol = 1e-10,
  ...
)
}
\arguments{
\item{formula.H0}{a two-sided linear formula  under the null, using the syntax of the \code{lme4} package.}

\item{data}{a data frame (required) containing the variables named in formula.}

\item{K}{the kernel matrix, which quantifies the similarities between samples.}

\item{relmat}{an optional correlation structure, see \code{lme4qtl} package.}

\item{family}{a GLM family, see \code{glm} and \code{family}.}

\item{acc}{error bound for \code{davies} method.}

\item{lim}{maximum number of integration terms for \code{davies} method.}

\item{tol}{the eigenvalue cutoff, below which it is considered to be 0. This is used to reduce the computation burden.}

\item{...}{arguments passing to mixed model \code{glmer} and \code{relmatGlmer}.}
}
\value{
A list containing
\describe{
  \item{p.value}{association p-value}
  \item{Q.adj}{ adjusted score statistic}
}
}
\description{
Compute the adjusted score statistic and p-value for GLMM
}
\examples{
set.seed(123)
m <- 20; k <- 4; n <- m * k; s <- 20
x <- rnorm(n); z <- matrix(sample(0 : 1, n * s, repl = TRUE, prob = c(0.8, 0.2)), n, s)
a <- rnorm(s); b <- 1
y <- rbinom(n, 1, prob = binomial()$linkinv(z \%*\% a + x * b + 0.5 * rnorm(m)[gl(m, k)]))
data <- data.frame(y = as.numeric(y), x = x, subject = gl(m, k))
rownames(data) <- paste(1:n)
K <- z \%*\% t(z)
VSAT(formula.H0 = y ~ x + (1 | subject), data = data, K = K, family = binomial()) 
VSAT(formula.H0 = y ~ (1 | subject), data = data, K = K, family = binomial()) # No covariate
}
\references{
Zhan X, et al. (2020)  Variant-Set Association Test for Generalized Linear Mixed Model.Genet Epidemiol. Submitted.
}
\author{
Jun Chen
}
\keyword{SKAT}
